Switch Location coordinates to BigDecimal for precision

+++ modules/common/src/main/java/com/quiptmc2/minecraft/config/objects/ConfigLocation.java
@@ -7,11 +7,13 @@ import com.quiptmc2.core.data.annotations.NotNull;
 import com.quiptmc2.minecraft.location.Location;
 import org.json.JSONObject;
 
+import java.math.BigDecimal;
+
 public class ConfigLocation extends ConfigObject {
 
 
-    public double x, y, z;
-    public float yaw, pitch;
+    public BigDecimal x, y, z;
+    public BigDecimal yaw, pitch;
     public String world;
 
     public ConfigLocation(QuiptIntegration integration, String id, Location location) {
@@ -37,21 +39,21 @@ public class ConfigLocation extends ConfigObject {
 
     public ConfigLocation(QuiptIntegration integration, String id, int blockX, int blockY, int blockZ, float yaw, float pitch, @NotNull String world) {
         this(integration, id);
-        this.x = blockX;
-        this.y = blockY;
-        this.z = blockZ;
-        this.yaw = yaw;
-        this.pitch = pitch;
+        this.x = BigDecimal.valueOf(blockX);
+        this.y = BigDecimal.valueOf(blockY);
+        this.z = BigDecimal.valueOf(blockZ);
+        this.yaw = BigDecimal.valueOf(yaw);
+        this.pitch = BigDecimal.valueOf(pitch);
         this.world = world;
     }
 
     public ConfigLocation(QuiptIntegration integration, String id, double x, double y, double z, float yaw, float pitch, @NotNull String world) {
         this(integration, id);
-        this.x = x;
-        this.y = y;
-        this.z = z;
-        this.yaw = yaw;
-        this.pitch = pitch;
+        this.x = BigDecimal.valueOf(x);
+        this.y = BigDecimal.valueOf(y);
+        this.z = BigDecimal.valueOf(z);
+        this.yaw = BigDecimal.valueOf(yaw);
+        this.pitch = BigDecimal.valueOf(pitch);
         this.world = world;
     }
